refactor: move data and scoring to modules
