+++ src/data/followup.js
@@ -1,0 +1,53 @@
+export const followupQuestions = {
+    devotee_pure: [
+      'Do you often feel a deep longing for union with the divine?',
+      'Does devotional music or chanting move you profoundly?',
+      'Do you easily offer your successes and failures to a higher power?',
+      'Does your heart expand during prayer or meditation?',
+      'Do you see love as the ultimate path to liberation?'
+    ],
+    devotee_doer: [
+      'Is serving others your primary expression of devotion?',
+      'Do you feel energized when acting on behalf of a noble cause?',
+      'Do you enjoy organizing service projects or community events?',
+      'Do you balance personal ambition with a sense of spiritual duty?',
+      'Do you often feel guided in your actions by divine inspiration?'
+    ],
+    devotee_knowledge: [
+      'Do you study spiritual texts to deepen your devotion?',
+      'Do insights into reality open your heart even more?',
+      'Do you prefer contemplation over rituals?',
+      'Does understanding spiritual principles motivate your service?',
+      'Is wisdom inseparable from love in your experience?'
+    ],
+    knowledge_doer: [
+      'Do you apply philosophical concepts directly in your life?',
+      'Does acting from clarity feel more important than emotional impulse?',
+      'Do you value discipline as a way to embody wisdom?',
+      'Do you seek mentors to refine your skills?',
+      'Is living your truth the highest form of knowledge?'
+    ],
+    pure_doer: [
+      'Do you act without expecting anything in return?',
+      'Does work feel natural when it benefits others?',
+      'Do you prefer direct action over introspection?',
+      'Do you feel most fulfilled when fully immersed in tasks?',
+      'Is detachment from results easy for you?'
+    ],
+    pure_knowledge: [
+      'Do you frequently inquire into "Who am I?"',
+      'Is inner silence more valuable than external achievements?',
+      'Do you naturally witness thoughts without identifying with them?',
+      'Does the idea of non-duality resonate deeply with you?',
+      'Do you feel that true knowledge cannot be expressed in words?'
+    ]
+  };
+
+export const followupInsights = {
+    devotee_pure: 'Your devotion is deep and heartfelt. Continue surrendering.',
+    devotee_doer: 'Service further opens your heart. Keep helping others.',
+    devotee_knowledge: 'Blend your love with wisdom for balanced growth.',
+    knowledge_doer: 'Act from clarity and let wisdom guide your deeds.',
+    pure_doer: 'Selfless action is your greatest teacher. Trust the flow.',
+    pure_knowledge: 'Abide as awareness and let silence reveal the truth.'
+  };
