+++ src/data/guidance.js
@@ -1,0 +1,32 @@
+export const gitaGuidance = {
+    devotee_pure: {
+      verse: 'BG 9.22',
+      text: 'To those who are constantly devoted and worship Me with love, I give the understanding by which they can come to Me.',
+      guidance: 'Your path is pure devotion. Cultivate love and surrender.'
+    },
+    devotee_doer: {
+      verse: 'BG 3.9',
+      text: 'Work done as a sacrifice for the Supreme must be performed, otherwise work binds one to this material world.',
+      guidance: 'Let every action be an offering. Serve selflessly.'
+    },
+    devotee_knowledge: {
+      verse: 'BG 7.17',
+      text: 'Of these, the wise one who is in full knowledge in union with me through devotion is the best.',
+      guidance: 'Let wisdom illuminate your love.'
+    },
+    knowledge_doer: {
+      verse: 'BG 4.18',
+      text: 'One who sees inaction in action, and action in inaction, is intelligent among men.',
+      guidance: 'Act from clear understanding.'
+    },
+    pure_doer: {
+      verse: 'BG 2.47',
+      text: 'You have a right to perform your prescribed duty, but you are not entitled to the fruits of action.',
+      guidance: 'Focus on perfect action without attachment.'
+    },
+    pure_knowledge: {
+      verse: 'BG 4.38',
+      text: 'In this world, there is nothing so sublime and pure as transcendental knowledge.',
+      guidance: 'Seek the highest truth through inquiry.'
+    }
+  };
