+++ src/data/pdfs.js
@@ -1,0 +1,9 @@
+export const personalityPDFs = {
+    devotee_pure: 'pdfs/devotee_pure.pdf',
+    devotee_doer: 'pdfs/devotee_doer.pdf',
+    devotee_knowledge: 'pdfs/devotee_knowledge.pdf',
+    knowledge_doer: 'pdfs/knowledge_doer.pdf',
+    pure_doer: 'pdfs/pure_doer.pdf',
+    pure_knowledge: 'pdfs/pure_knowledge.pdf'
+  };
+
