+++ src/data/personality.js
@@ -1,0 +1,108 @@
+export const personalityTypes = {
+    devotee_pure: {
+      name: 'The Pure Devotee',
+      subtitle: 'Lover of Divine',
+      description: 'You embody pure divine love and complete surrender to the highest truth.',
+      emoji: '🙏',
+      practices: [
+        'How can I dissolve completely in divine love today?',
+        'What would pure surrender look like in this moment?',
+        'How do I serve without any sense of doership?',
+        'What is blocking my heart from opening fully?',
+        'How can I see only the Divine in everything?'
+      ]
+    },
+    devotee_doer: {
+      name: 'The Devotee-Doer',
+      subtitle: 'Divine Servant in Action',
+      description: 'You express divine love through dedicated service and righteous action.',
+      emoji: '💝',
+      practices: [
+        'How can I serve the Divine through my work today?',
+        'What actions align with divine will?',
+        'How do I offer all results to the Divine?',
+        'Where is service needed most today?',
+        'How can I work without attachment to outcomes?'
+      ]
+    },
+    devotee_knowledge: {
+      name: 'The Devotee-Sage',
+      subtitle: 'Divine Lover-Knower',
+      description: 'You unite divine love with sacred wisdom and deep understanding.',
+      emoji: '🕊️',
+      practices: [
+        'What is the Divine trying to teach me today?',
+        'How does love and wisdom unite in my heart?',
+        'What is the nature of the Divine Self?',
+        'How do I know with the heart, not just the mind?',
+        'What illusions is love dissolving?'
+      ]
+    },
+    knowledge_doer: {
+      name: 'The Sage-Warrior',
+      subtitle: 'Wise Actor',
+      description: 'You combine deep wisdom with purposeful action in service of truth.',
+      emoji: '⚔️',
+      practices: [
+        'How do I act from wisdom rather than reactivity?',
+        'What actions serve the highest truth today?',
+        'How do I remain unattached while fully engaged?',
+        'What is my dharmic duty in this situation?',
+        'Where can wisdom be applied practically?'
+      ]
+    },
+    pure_doer: {
+      name: 'The Pure Karma Yogi',
+      subtitle: 'Selfless Actor',
+      description: 'You embody perfect action without attachment.',
+      emoji: '🌟',
+      practices: [
+        'How do I act without any sense of doership?',
+        'What would perfect selfless service look like?',
+        'How do I work without attachment to results?',
+        'Where can I serve most effectively today?',
+        'How do I become a pure instrument?'
+      ]
+    },
+    pure_knowledge: {
+      name: 'The Pure Jnani',
+      subtitle: 'Knower of Truth',
+      description: 'You embody pure wisdom and direct knowledge.',
+      emoji: '🧘',
+      practices: [
+        'Who am I beyond all identifications?',
+        'What is the nature of ultimate reality?',
+        'How do I rest in pure awareness?',
+        'What is real and what is illusion?',
+        'How do I abide in the Self?'
+      ]
+    } 
+  };
+
+export const personalityInsights = {
+    devotee_pure: [
+      'Let devotion be your guide in every moment.',
+      'Surrender fully to what is and notice the peace that follows.'
+    ],
+    devotee_doer: [
+      'Channel your love into acts of service today.',
+      'Small acts done with great love transform the world.'
+    ],
+    devotee_knowledge: [
+      'Let wisdom soften your heart into deeper compassion.',
+      'Contemplate how love and truth arise together.'
+    ],
+    knowledge_doer: [
+      'Apply your insights in practical ways for others.',
+      'Wise action is the bridge between understanding and service.'
+    ],
+    pure_doer: [
+      'Selfless work purifies the mind and opens the heart.',
+      'See if you can act without clinging to the outcome.'
+    ],
+    pure_knowledge: [
+      'Rest in awareness and notice the quiet joy within.',
+      'Let each experience reveal the truth of who you are.'
+    ]
+  };
+
