+++ src/data/questions.js
@@ -1,0 +1,32 @@
+export const questions = [
+    { id: 1, category: 'Devotion', text: "Do you sometimes feel a longing you can't quite name or explain?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 4, category: 'Devotion', text: 'Do certain moments of beauty or kindness bring tears to your eyes?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 7, category: 'Devotion', text: 'Do you find peace when you stop trying to control outcomes?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 10, category: 'Devotion', text: "Have you ever felt the boundary between 'you' and 'everything else' dissolve?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 13, category: 'Devotion', text: 'Does seeing others suffer genuinely hurt your heart?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 16, category: 'Devotion', text: "Do you sometimes feel love so strong it's almost too much to contain?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 19, category: 'Devotion', text: "When someone calls you spiritual, do you feel they don't really know you?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 22, category: 'Devotion', text: 'Do beautiful moments feel like unexpected gifts?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 25, category: 'Devotion', text: 'Do you sometimes see something precious in every person you meet?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 28, category: 'Devotion', text: 'Do you ever feel overwhelming gratitude for existence itself?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 2, category: 'Knowledge', text: "Do you sometimes wonder if what you think is 'you' is actually real?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 5, category: 'Knowledge', text: 'Do thoughts sometimes feel like clouds passing through an open sky?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 8, category: 'Knowledge', text: "Do you find yourself wondering 'What is aware of my thoughts?'", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 12, category: 'Knowledge', text: 'Do you feel drawn to understand the deepest truths about existence?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 15, category: 'Knowledge', text: 'Are you sometimes aware of an unchanging presence that watches all experience?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 18, category: 'Knowledge', text: 'Do you sometimes question the very nature of reality itself?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 21, category: 'Knowledge', text: "Do you feel that what you call 'mind' is actually empty space?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 24, category: 'Knowledge', text: 'Do you feel more real in moments of complete stillness?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 27, category: 'Knowledge', text: "Have you ever realized that the 'searcher' and the 'sought' are the same?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 30, category: 'Knowledge', text: 'Do you sometimes feel that all knowledge is actually remembering?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 3, category: 'Action', text: 'When you see something that needs doing, do you naturally step forward?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 6, category: 'Action', text: 'Do people naturally come to you when things get difficult?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 9, category: 'Action', text: "Do you sometimes act so naturally that you forget you're the one doing it?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 11, category: 'Action', text: "Do you sometimes act to help others without thinking what's in it for you?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 14, category: 'Action', text: "Do you sometimes give or serve so naturally you don't think about it?", options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 17, category: 'Action', text: 'Do you sense when your actions flow with something greater than your personal will?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 20, category: 'Action', text: 'Do you naturally take charge when situations become chaotic?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 23, category: 'Action', text: 'Can you work with full intensity while feeling completely detached from results?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 26, category: 'Action', text: 'Do you prefer taking decisive action over endless analysis?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] },
+    { id: 29, category: 'Action', text: 'Do you feel most alive when serving others or a greater cause?', options: ['Never', 'Rarely', 'Sometimes', 'Often', 'Always'] }
+  ];
